requirements-notebook.txt
jupyterlab
pandas
numpy
scikit-learn
matplotlib
seaborn
bertopic
sentence-transformers
umap-learn
hdbscan
python-dotenv
